--- lang/example-txtx/eval.py
@@ -1,2 +1,0 @@
-def eval_statement(statement):
-    print(statement.lit)
