--- lang/example-txtx/grammar.tx
@@ -1,23 +1,0 @@
-Grammar:
-  statements*=Statement | Comment | Skip
-;
-
-Statement:
-  'echo' lit=Literal StatementEnd
-;
-
-Literal:
-  NUMBER | STRING
-;
-
-StatementEnd:
-  ';'
-;
-
-Comment:
-  /\/\/.*\n/
-;
-
-Skip:
-  /[\n\t\r]/
-;
