+++ lang/src/spm/spm/file_utils.py
@@ -3,24 +3,16 @@ from os import path
 
 
 def load_local_packages():
-
     with open("package.json", "r") as fp:
-        packages = fp.read()
-
-        return json.loads(packages)["packages"]
+        return json.load(fp)["packages"]
 
 def load_solidity_file(file_path):
     with open(file_path.replace('"', ''), "r") as fp:
-        input = fp.read()
-
-        return input
+        return fp.read()
 
 def load_package(package_name, package_version):
-    with open("{0}_{1}.json".format(package_name, package_version), "r") as fp:
-        package = fp.read()
-
-        return json.loads(package)
-
+    with open(f"{package_name}_{package_version}.json", "r") as fp:
+        return json.load(fp)
 
 def check_file_path(filePath):
     if not path.isfile(filePath.replace('"', '')):
